Adds clustering in node server; adds links to header

+++ server/server.js
@@ -1,49 +1,31 @@
-const path = require('path');
-const express = require('express');
-const morgan = require('morgan');
-const bodyParser = require('body-parser');
-const helmet = require('helmet');
-const passport = require('passport');
-const compression = require('compression');
+/* eslint-disable no-console */
+const cluster = require('cluster');
 
-if (process.env.NODE_ENV !== 'production') {
-  require('./config/env');
+if (cluster.isMaster) {
+  const cores = require('os').cpus();
+
+  console.log('Master cluster setting up ' + cores.length + ' workers...');
+
+  cores.forEach(() => {
+    cluster.fork();
+  });
+
+  cluster.on('online', worker => {
+    console.log('Worker ' + worker.process.pid + ' is online');
+  });
+
+  cluster.on('exit', (worker, code, signal) => {
+    console.log('Worker ' + worker.process.pid + ' died with code: ' + code + ', and signal: ' + signal);
+    console.log('Starting a new worker');
+    cluster.fork();
+  });
+} else {
+  const app = require('./app');
+  const models = require('./models');
+
+  models.sequelize.sync({ force: false }).then(() => {
+    app.listen(process.env.PORT, () => console.log(`Server listening on port ${process.env.PORT}`));
+  });
 }
-const models = require('./models');
-const routes = require('./routes');
-const redisSession = require('./config/session');
 
-const app = express();
-const publicPath = path.join(__dirname, '../public/dist');
-
-/*
-MIDDLEWARE
-*/
-if (process.env.NODE_ENV !== 'test') {
-  app.use(morgan('dev'));
-}
-
-app.use(compression());
-app.use(express.static(publicPath));
-app.use(helmet());
-app.use(bodyParser.json({ type: '*/*' }));
-app.use(redisSession);
-
-// Init passport session
-app.use(passport.initialize());
-app.use(passport.session());
-
-/*
-ROUTES
-*/
-app.use('/api', routes);
-app.get('*', (req, res) => res.sendFile('index.html', { root: publicPath }));
-
-/*
-INIT DB + RUN SERVER
-*/
-models.sequelize.sync({ force: false }).then(() => {
-  app.listen(process.env.PORT, () => console.log(`Server listening on port ${process.env.PORT}`));  // eslint-disable-line no-console
-});
-
-module.exports = app;
+/* eslint-disable no-console */
